requirements.txt
requests
praw
python-dotenv
pytest
pylint
